+++ src/test/java/com/github/alexdlaird/ngrok/NgrokClientTest.java
@@ -96,7 +96,7 @@ class NgrokClientTest extends NgrokTestCase {
         assertFalse(ngrokProcessV2.isRunning());
         final CreateTunnel createTunnel = new CreateTunnel.Builder()
                 .withAddr(5000)
-                .build();
+                .build(javaNgrokConfigV2);
 
         // WHEN
         final Tunnel tunnel = ngrokClientV2.connect(createTunnel);
@@ -129,7 +129,7 @@ class NgrokClientTest extends NgrokTestCase {
         assertFalse(ngrokClientV3.getNgrokProcess().isRunning());
         final CreateTunnel createTunnel = new CreateTunnel.Builder()
                 .withAddr(5000)
-                .build();
+                .build(javaNgrokConfigV3);
 
         // WHEN
         final Tunnel tunnel = ngrokClientV3.connect(createTunnel);
@@ -161,7 +161,7 @@ class NgrokClientTest extends NgrokTestCase {
         // GIVEN
         final CreateTunnel createTunnel = new CreateTunnel.Builder()
                 .withName("my-tunnel")
-                .build();
+                .build(javaNgrokConfigV2);
 
         // WHEN
         final Tunnel tunnel = ngrokClientV2.connect(createTunnel);
@@ -175,9 +175,9 @@ class NgrokClientTest extends NgrokTestCase {
     @Test
     public void testMultipleConnectionsNoTokenFailsV2() throws InterruptedException {
         // WHEN
-        ngrokClientV2.connect(new CreateTunnel.Builder().withAddr(5000).build());
+        ngrokClientV2.connect(new CreateTunnel.Builder().withAddr(5000).build(javaNgrokConfigV2));
         Thread.sleep(1000);
-        final JavaNgrokHTTPException exception = assertThrows(JavaNgrokHTTPException.class, () -> ngrokClientV2.connect(new CreateTunnel.Builder().withAddr(5001).build()));
+        final JavaNgrokHTTPException exception = assertThrows(JavaNgrokHTTPException.class, () -> ngrokClientV2.connect(new CreateTunnel.Builder().withAddr(5001).build(javaNgrokConfigV2)));
 
         // THEN
         assertEquals(HTTP_BAD_GATEWAY, exception.getStatusCode());
@@ -214,7 +214,7 @@ class NgrokClientTest extends NgrokTestCase {
         // GIVEN
         final CreateTunnel createTunnel = new CreateTunnel.Builder()
                 .withBindTls(BindTls.BOTH)
-                .build();
+                .build(javaNgrokConfigV2);
         final Tunnel tunnel = ngrokClientV2.connect(createTunnel);
 
         // WHEN
@@ -262,7 +262,7 @@ class NgrokClientTest extends NgrokTestCase {
         // GIVEN
         final CreateTunnel createTunnel = new CreateTunnel.Builder()
                 .withSchemes(List.of("http"))
-                .build();
+                .build(javaNgrokConfigV3);
         final Tunnel tunnel = ngrokClientV3.connect(createTunnel);
 
         // WHEN
@@ -278,7 +278,7 @@ class NgrokClientTest extends NgrokTestCase {
         // GIVEN
         final CreateTunnel createTunnel = new CreateTunnel.Builder()
                 .withSchemes(List.of("https"))
-                .build();
+                .build(javaNgrokConfigV3);
         final Tunnel tunnel = ngrokClientV3.connect(createTunnel);
 
         // WHEN
@@ -294,7 +294,7 @@ class NgrokClientTest extends NgrokTestCase {
         // GIVEN
         final CreateTunnel createTunnel = new CreateTunnel.Builder()
                 .withSchemes(List.of("http", "https"))
-                .build();
+                .build(javaNgrokConfigV3);
         final Tunnel tunnel = ngrokClientV3.connect(createTunnel);
 
         // WHEN
@@ -310,7 +310,7 @@ class NgrokClientTest extends NgrokTestCase {
         // GIVEN
         final CreateTunnel createTunnel = new CreateTunnel.Builder()
                 .withBindTls(BindTls.BOTH)
-                .build();
+                .build(javaNgrokConfigV3);
         final Tunnel tunnel = ngrokClientV3.connect(createTunnel);
 
         // WHEN
@@ -326,7 +326,7 @@ class NgrokClientTest extends NgrokTestCase {
         // GIVEN
         final CreateTunnel createTunnel = new CreateTunnel.Builder()
                 .withName("my-tunnel")
-                .build();
+                .build(javaNgrokConfigV2);
         final Tunnel tunnel = ngrokClientV2.connect(createTunnel);
         assertTrue(ngrokClientV2.getNgrokProcess().isRunning());
         final List<Tunnel> tunnels1 = ngrokClientV2.getTunnels();
@@ -345,7 +345,7 @@ class NgrokClientTest extends NgrokTestCase {
         // GIVEN
         final CreateTunnel createTunnel = new CreateTunnel.Builder()
                 .withName("my-tunnel")
-                .build();
+                .build(javaNgrokConfigV3);
         final Tunnel tunnel = ngrokClientV3.connect(createTunnel);
         assertTrue(ngrokClientV3.getNgrokProcess().isRunning());
 
@@ -360,7 +360,7 @@ class NgrokClientTest extends NgrokTestCase {
     @Test
     public void testKill() {
         // GIVEN
-        final CreateTunnel createTunnel = new CreateTunnel.Builder().withName("my-tunnel").build();
+        final CreateTunnel createTunnel = new CreateTunnel.Builder().withName("my-tunnel").build(javaNgrokConfigV2);
         ngrokClientV2.connect(createTunnel);
 
         // WHEN
@@ -399,7 +399,7 @@ class NgrokClientTest extends NgrokTestCase {
         final CreateTunnel createTunnel = new CreateTunnel.Builder()
                 .withProto(Proto.TCP)
                 .withAddr(5000)
-                .build();
+                .build(javaNgrokConfigV2);
 
         // WHEN
         final Tunnel tunnel = ngrokClient2.connect(createTunnel);
@@ -431,7 +431,7 @@ class NgrokClientTest extends NgrokTestCase {
         final CreateTunnel createTunnel = new CreateTunnel.Builder()
                 .withProto(Proto.TCP)
                 .withAddr(5000)
-                .build();
+                .build(javaNgrokConfigV2);
 
         // WHEN
         final Tunnel tunnel = ngrokClient2.connect(createTunnel);
@@ -463,7 +463,7 @@ class NgrokClientTest extends NgrokTestCase {
         final String subdomain = createUniqueSubdomain();
         final CreateTunnel createTunnel = new CreateTunnel.Builder()
                 .withSubdomain(subdomain)
-                .build();
+                .build(javaNgrokConfigV2);
 
         // WHEN
         final Tunnel tunnel = ngrokClient2.connect(createTunnel);
@@ -486,7 +486,7 @@ class NgrokClientTest extends NgrokTestCase {
         assertFalse(ngrokClientV2.getNgrokProcess().isRunning());
         final CreateTunnel createTunnel = new CreateTunnel.Builder()
                 .withAddr("file:///")
-                .build();
+                .build(javaNgrokConfigV2);
 
         // WHEN
         final Tunnel tunnel = ngrokClientV2.connect(createTunnel);
@@ -510,7 +510,7 @@ class NgrokClientTest extends NgrokTestCase {
         assertFalse(ngrokClientV2.getNgrokProcess().isRunning());
         final CreateTunnel createTunnel = new CreateTunnel.Builder()
                 .withAddr("file:///")
-                .build();
+                .build(javaNgrokConfigV2);
         final String publicUrl = ngrokClientV2.connect(createTunnel).getPublicUrl();
         Thread.sleep(1000);
 
@@ -535,7 +535,7 @@ class NgrokClientTest extends NgrokTestCase {
         assertFalse(ngrokClientV2.getNgrokProcess().isRunning());
         final CreateTunnel createTunnel = new CreateTunnel.Builder()
                 .withAddr("file:///")
-                .build();
+                .build(javaNgrokConfigV2);
         final Tunnel tunnel = ngrokClientV2.connect(createTunnel);
         Thread.sleep(1000);
         final String apiUrl = ngrokClientV2.getNgrokProcess().getApiUrl();
@@ -606,11 +606,11 @@ class NgrokClientTest extends NgrokTestCase {
         // WHEN
         final CreateTunnel createHttpTunnel = new CreateTunnel.Builder()
                 .withName("http-tunnel")
-                .build();
+                .build(javaNgrokConfigV2);
         final Tunnel httpTunnel = ngrokClient2.connect(createHttpTunnel);
         final CreateTunnel createSshTunnel = new CreateTunnel.Builder()
                 .withName("tcp-tunnel")
-                .build();
+                .build(javaNgrokConfigV2);
         final Tunnel sshTunnel = ngrokClient2.connect(createSshTunnel);
         final List<Tunnel> tunnels = ngrokClient2.getTunnels();
 
@@ -663,11 +663,11 @@ class NgrokClientTest extends NgrokTestCase {
         // WHEN
         final CreateTunnel createHttpTunnel = new CreateTunnel.Builder()
                 .withName("http-tunnel")
-                .build();
+                .build(javaNgrokConfigV3);
         final Tunnel httpTunnel = ngrokClient2.connect(createHttpTunnel);
         final CreateTunnel createSshTunnel = new CreateTunnel.Builder()
                 .withName("tcp-tunnel")
-                .build();
+                .build(javaNgrokConfigV3);
         final Tunnel sshTunnel = ngrokClient2.connect(createSshTunnel);
         final List<Tunnel> tunnels = ngrokClient2.getTunnels();
 
@@ -713,7 +713,7 @@ class NgrokClientTest extends NgrokTestCase {
         final CreateTunnel createTunnelSubdomain = new CreateTunnel.Builder()
                 .withSubdomain(subdomain2)
                 .withAddr(5000)
-                .build();
+                .build(javaNgrokConfigV2);
         final Tunnel ngrokTunnel2 = ngrokClient2.connect(createTunnelSubdomain);
 
         // THEN